+++ ja2lib/Soldier.h
@@ -13,8 +13,8 @@ struct SOLDIERTYPE *GetSoldierByID(int index);
 u8 GetSolID(const struct SOLDIERTYPE *s);
 u8 GetSolProfile(const struct SOLDIERTYPE *s);
 
-i16 GetSolSectorX(const struct SOLDIERTYPE *s);
-i16 GetSolSectorY(const struct SOLDIERTYPE *s);
+u8 GetSolSectorX(const struct SOLDIERTYPE *s);
+u8 GetSolSectorY(const struct SOLDIERTYPE *s);
 i8 GetSolSectorZ(const struct SOLDIERTYPE *s);
 SectorID8 GetSolSectorID8(const struct SOLDIERTYPE *s);
 SectorID16 GetSolSectorID16(const struct SOLDIERTYPE *s);
